Create script to generate React component
This commit adds a new script, create-component.js, which can be used to generate a new React component. The script takes a component name as an argument and creates the necessary files and directories for the component, including the component file itself, a model file, a type file, and a styles file. It also creates index files for the components and hooks directories.

The script performs some validations, such as checking if the component name is in CamelCase and if a component with the same name already exists.

This script will be useful for quickly generating new components in the project.

Fix package.json script

This commit fixes a typo in the package.json file. The "prepare" script was missing a comma at the end, causing a syntax error. The missing comma has been added.

The "prepare" script is used by the husky package to set up Git hooks.